[contrakit] Improve project setup
Move tests inside src folder.
This simplifies the project setup, having only one tsconfig.json
and thus test that don't compile will be catcher earlier.

npmignore and package.json were changed so not to include test files in
npm package.
