+++ packages/contractkit/.npmignore
@@ -1,8 +1,14 @@
-/node_modules
-/coverage
+/.devchain/
+/coverage/
+/node_modules/
+/src/
+
 /tslint.json
-/tsconfig.json
-/test
+/tsconfig.*
+/jest.config.*
+*.tgz
+
 /src
-/jest.config.json
-/.devchain
+
+/lib/**/*.test.*
+/lib/test-utils
