+++ packages/contractkit/jest.config.js
@@ -1,18 +1,8 @@
-const { pathsToModuleNameMapper } = require('ts-jest/utils')
-
-const { compilerOptions } = require('./test/tsconfig')
-
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, { prefix: '<rootDir>/' }),
-  globals: {
-    'ts-jest': {
-      tsConfig: '<rootDir>/test/tsconfig.json',
-    },
-  },
-  testPathIgnorePatterns: ['<rootDir>/lib/', '<rootDir>/node_modules/'],
-  setupFilesAfterEnv: ['<rootDir>/test/test-utils/matchers.ts'],
-  globalSetup: '<rootDir>/test/test-utils/ganache.setup.ts',
-  globalTeardown: '<rootDir>/test/test-utils/ganache.teardown.ts',
+  testMatch: ['<rootDir>/src/**/?(*.)+(spec|test).ts?(x)'],
+  setupFilesAfterEnv: ['<rootDir>/src/test-utils/matchers.ts'],
+  globalSetup: '<rootDir>/src/test-utils/ganache.setup.ts',
+  globalTeardown: '<rootDir>/src/test-utils/ganache.teardown.ts',
 }
