+++ packages/contractkit/package.json
@@ -40,8 +40,5 @@
   },
   "engines": {
     "node": ">=8.13.0"
-  },
-  "files": [
-    "lib/**/*"
-  ]
+  }
 }
