+++ packages/contractkit/src/contract-cache.test.ts
@@ -1,18 +1,26 @@
-import { AllContracts, newKit } from '@src/index'
-import { Web3ContractCache } from '@src/web3-contract-cache'
+import { CeloContract, newKit } from '.'
+import { ValidWrappers, WrapperCache } from './contract-cache'
 
-function newWeb3ContractCache() {
+const TestedWrappers: ValidWrappers[] = [
+  CeloContract.GoldToken,
+  CeloContract.StableToken,
+  CeloContract.Exchange,
+  CeloContract.Validators,
+  CeloContract.LockedGold,
+]
+
+function newWrapperCache() {
   const kit = newKit('')
   const AnyContractAddress = '0xe832065fb5117dbddcb566ff7dc4340999583e38'
   jest.spyOn(kit.registry, 'addressFor').mockResolvedValue(AnyContractAddress)
-  const contractCache = new Web3ContractCache(kit)
+  const contractCache = new WrapperCache(kit)
   return contractCache
 }
 
 describe('getContract()', () => {
-  const contractCache = newWeb3ContractCache()
+  const contractCache = newWrapperCache()
 
-  for (const contractName of AllContracts) {
+  for (const contractName of TestedWrappers) {
     test(`SBAT get ${contractName}`, async () => {
       const contract = await contractCache.getContract(contractName)
       expect(contract).not.toBeNull()
@@ -22,8 +30,8 @@ describe('getContract()', () => {
 })
 
 test('should cache contracts', async () => {
-  const contractCache = newWeb3ContractCache()
-  for (const contractName of AllContracts) {
+  const contractCache = newWrapperCache()
+  for (const contractName of TestedWrappers) {
     const contract = await contractCache.getContract(contractName)
     const contractBis = await contractCache.getContract(contractName)
     expect(contract).toBe(contractBis)
