+++ packages/contractkit/src/test-utils/ganache-test.ts
@@ -1,6 +1,6 @@
-import { injectDebugProvider } from '@src/providers/debug-provider'
 import Web3 from 'web3'
 import { JsonRPCResponse } from 'web3/providers'
+import { injectDebugProvider } from '../providers/debug-provider'
 
 export function jsonRpcCall<O>(web3: Web3, method: string, params: any[]): Promise<O> {
   return new Promise<O>((resolve, reject) => {
