+++ packages/contractkit/src/test-utils/ganache.teardown.ts
@@ -1,6 +1,6 @@
 export default function tearDown() {
   console.log('Stopping ganache')
-  ;(global as any).stopGanache().catch((err: any) => {
+  return (global as any).stopGanache().catch((err: any) => {
     console.error('error stopping ganache')
     console.error(err)
   })
