+++ packages/contractkit/src/utils/tx-result.test.ts
@@ -1,5 +1,5 @@
-import { toTxResult } from '@src/utils/tx-result'
 import { promiEventSpy } from '../test-utils/PromiEventStub'
+import { toTxResult } from './tx-result'
 
 test('should resolve to hash & receipt on normal flow', async () => {
   const peStub = promiEventSpy<number>()
