--- packages/contractkit/test/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../typescript/tsconfig.library.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "noEmit": true,
-    "baseUrl": "..",
-    "paths": {
-      "@src/*": ["src/*"]
-    }
-  },
-  "include": ["."],
-  "references": [{ "path": ".." }]
-}
